Add thread locks around TTLCache access
The TTLCache provided by cachetools is not threadsafe. We have seen
cases in the wild where the caching code calls a method to delete a
token from the cache and then gets a `KeyError` on the `del` call:

Cachetools version: 4.2.2 (out of date, but I wanted to keep the changes
here minimal)
Example exception:
```
Traceback (most recent call last):
  File "/kb/module/lib/AbstractHandle/AbstractHandleServer.py", line
101, in _call_method
    result = method(ctx, *params)
  File "/kb/module/lib/AbstractHandle/AbstractHandleImpl.py", line 335,
in add_read_acl
    returnVal = self.handler.add_read_acl(hids, ctx['token'],
username=username)
  File "/kb/module/lib/AbstractHandle/Utils/Handler.py", line 232, in
add_read_acl
    if not self._is_admin_user(token):
  File "/kb/module/lib/AbstractHandle/Utils/Handler.py", line 102, in
_is_admin_user
    self.token_cache[token] = {'customroles': customroles}
  File "/miniconda/lib/python3.6/site-packages/cachetools/ttl.py", line
87, in __setitem__
    self.expire(time)
  File "/miniconda/lib/python3.6/site-packages/cachetools/ttl.py", line
167, in expire
    cache_delitem(self, curr.key)
  File "/miniconda/lib/python3.6/site-packages/cachetools/cache.py",
line 66, in __delitem__
    del self.__data[key]
KeyError: '<token removed>'
```

The assumption here is that two threads are calling the expire() method
at the same time, one deletes the token, and the other has nothing to
delete and fails. This is difficult to test ex-situ so the best plan is
probably to put develop up in prod for some amount of time and see if
the errors go away.

This PR also contains changes to the test suite and image build as the
prior image wouldn't build with a package-installed mongo. The changes
also make it easier to matrix test mongo in CI in the future.